\name{p1_H0}
\docType{data}
\alias{p1_H0}
\title{Signal of association of the statistic S1 under H0}
\description{
  This dataset contains the signal of association of the statistic S1 under the null hypothesis H0. It consists of 200 values, one for each simulations under H0. See the vignette for the definition of S1 and more details: \code{vignette("waffect-tutorial")}.
}
\usage{data(p1_H0)}
\format{A table with one column and 200 rows}
\keyword{dataset}