\name{p1_H1}
\docType{data}
\alias{p1_H1}
\title{Signal of association of the statistic S1 under H1}
\description{
  This dataset contains the signal of association of the statistic S1 under the alternative hypothesis H1. It consists of 200 values, one for each simulations under H1. See the vignette for the definition of S1 and more details: \code{vignette("waffect-tutorial")}.
}
\usage{data(p1_H1)}
\format{A table with one column and 200 rows}
\keyword{dataset}