\name{waffectbin}
\alias{waffectbin}
\docType{package}
\title{
Low level function for simulating binary phenotypes.
}
\description{
Usage of \code{waffectbin} is deprecated, please use \code{waffect}.
}
\seealso{
	Documentation for \code{\link{waffect}} and \code{\link{waffect-package}}. Useful information can be also found in the vignette: \code{vignette("waffect-tutorial")}. 
}